+++ changelog.md
@@ -1,3 +1,8 @@
+# 3.12.1
+
+-   Fixed a regression with `getEnv` sometimes not returning the proper environment.
+-   Fixed an issue where `map.put` would not work with snapshots of types with an optional id [#1131](https://github.com/mobxjs/mobx-state-tree/issues/1131) through [#1226](https://github.com/mobxjs/mobx-state-tree/pull/1226) by [@xaviergonz](https://github.com/xaviergonz).
+
 # 3.12.0
 
 -   Added `TypeOfValue<typeof variable>` to extract the type of a complex (non primitive) variable in Typescript.